Fix missing translation of editor open popover title
Bootstrap initialized the popover with the title from the button's title
attribute (which was then removed) but did not hook the popover into the
DOM. At that time of loading, the title attribute was not yet translated.

Now the bootstrap popover plugin is given a title property with a function
that is called on demand, when the translations are already loaded. The
button's title attribute is added later so that bootstrap does not remove it.